requirements.txt
celery
requests
bs4
pymongo[srv]
lxml
python-dateutil
python-slugify
tweepy
textblob
openai
pytrends
redis
slugify
google-api-python-client
google-auth
google-auth-oauthlib
isodate
youtube_transcript_api
opencv-python-headless
numpy
gtts
pyttsx3
pyobjc
python-dotenv
pytube
pytubefix
requests_oauthlib
oauth2client
